[FS-42][Server] Route and body validations for Stock and Comment

+++ server/controllers/StockController.cs
@@ -27,14 +27,16 @@ namespace server.controllers
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+      if(!ModelState.IsValid) return BadRequest(ModelState);
       var stocks = await _stockRepo.GetAllAsync();
       var stockDTO = stocks.Select(stock => stock.ToStockDTO());
       return Ok(stocks);
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:int}")]
     public async Task<IActionResult> GetById([FromRoute] int id)
     {
+      if(!ModelState.IsValid) return BadRequest(ModelState);
       var stock =await _stockRepo.GetByIdAsync(id);
       if (stock == null)
       {
@@ -46,16 +48,18 @@ namespace server.controllers
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateStockReqDTO newStock)
     {
+      if(!ModelState.IsValid) return BadRequest(ModelState);
       var stockModel = newStock.ToStockFromCreateDTO();
       await _stockRepo.CreateAsync(stockModel);
       return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDTO());
     }
 
     [HttpPut]
-    [Route("{id}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockReqDTO updatedStock)
     {
-      var stockModel = await _stockRepo.UpdateAsync(id, updatedStock);
+      if(!ModelState.IsValid) return BadRequest(ModelState);
+      var stockModel = await _stockRepo.UpdateAsync(id, updatedStock.ToStockFromUpdateDTO());
       if (stockModel == null)
       {
         return NotFound();
@@ -64,9 +68,10 @@ namespace server.controllers
     }
 
     [HttpDelete]
-    [Route("{id}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Delete([FromRoute] int id)
     {
+      if(!ModelState.IsValid) return BadRequest(ModelState);
       var stockModel = await _stockRepo.DeleteAsync(id);
       if (stockModel == null)
       {
